% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_Visualise.R
\name{BatchEffect_PCAHeatmap_Merge}
\alias{BatchEffect_PCAHeatmap_Merge}
\title{Merge multiple heatmaps into one plot}
\usage{
BatchEffect_PCAHeatmap_Merge(x)
}
\arguments{
\item{x}{Epigenetic set object.}
}
\value{
Updated Epigenetic set object.
}
\description{
Merge multiple heatmaps into one plot
}
