% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_Visualise.R
\name{BatchEffect_PointPlot}
\alias{BatchEffect_PointPlot}
\title{Title}
\usage{
BatchEffect_PointPlot(
  x,
  batch,
  n = 6,
  xlim = c(-1000, 2500),
  ylim = c(-1000, 1000)
)
}
\arguments{
\item{ylim}{}
}
\value{

}
\description{
Title
}
