% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_prc.R
\name{BatchEffect_prc}
\alias{BatchEffect_prc}
\title{Title}
\usage{
BatchEffect_prc(
  x,
  n_pcas = 8,
  filename_info = c("Chip", "Well", "Swabnr", "Experiment_Method", "Patient_Group",
    "Patient_Number", "Experiment_Group", "Time"),
  filename_separator = "_"
)
}
\arguments{
\item{filename_separator}{}
}
\value{

}
\description{
Title
}
