% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_Visualise.R
\name{BatchEffect_Visualise}
\alias{BatchEffect_Visualise}
\title{BatchEffect_Visualise function}
\usage{
BatchEffect_Visualise(
  x,
  df,
  by,
  batch,
  Cat_vars = c(),
  Cont_vars = c(),
  n = 6,
  xlim = c(-1000, 2500),
  ylim = c(-1000, 1000)
)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{df}{Metadata}

\item{by}{Name of the column to merge existing data with}

\item{batch}{The non-biological factor that we want to investigate for batch effect.}

\item{Cat_vars}{Categorical variables for the heatmap}

\item{Cont_vars}{Continues variables for the heatmap}

\item{n}{number of samples per plot}

\item{xlim}{range of the x-axis}

\item{ylim}{range of the y-axis}
}
\value{
Updated Epigenetic set object.
}
\description{
Function to visualise batch effect using point plots and heatmaps. This function combines the MergeMetadata, BatchEffect_PointPlot and BatchEffect_PCAHeatmap function.
}
