% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMP_limma.R
\name{DMP_limma}
\alias{DMP_limma}
\title{Detect differentially methylated positions in the DNA using the limma package.}
\usage{
DMP_limma(set, Patient_info, cat_vars = NULL, cont_vars = NULL, Group = NULL)
}
\arguments{
\item{set}{Preprocessed methylation set, can be genomic methyl set or genomic ratio set.}

\item{Patient_info}{Data frame with adjustment variables.}

\item{cat_vars}{The variable names in Patient_info that are categorical variables}

\item{cont_vars}{The variable names in Patient_info that are continuous variables}

\item{Group}{The variable in Patient_info that represents which observations belong to which groups}
}
\value{
A list with limma output variables and temporary results
\itemize{
   \item DMPresult - Data table with statistics per CpG sites
   \item model - eBayes fitted linear model
   \item design - Design matrix
   \item contMatrix - Contrast matrix
   \item Sign_CpG - CpG names of DMPs
}
}
\description{
This function first uses the `lmfit` function from the `limma` package with methylset and the generated design matrix as its inputs.
Next it uses the `contrast.fit` function from the `limma` package with the `lmfit` output and the generated contrast matrix as its inputs.
The contrast matrix specifies the comparison of interest (e.g.: Patients vs Controls).
To compute a moderated t-statistic of differential expression, it uses empirical Bayes moderation of the standard error.
}
