% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_dmrcate.R
\name{DMR_dmrcate}
\alias{DMR_dmrcate}
\title{Detect differentially methylated positions in the DNA using the limma package.}
\usage{
DMR_dmrcate(set, DMP, lambda = 1000, fdr = 0.05)
}
\arguments{
\item{set}{Preprocessed methylation set, can be genomic methyl set or genomic ratio set.}

\item{DMP}{Output from `DMP_limma()` function.}

\item{lambda}{Gaussian kernel bandwidth for smoothed-function estimation. Also informs DMR bookend definition; gaps >= lambda between significant CpG sites will be in separate DMRs. Support is truncated at 5*lambda. Default is 1000 nucleotides.}

\item{fdr}{false discovery rate, default = 0.05.}
}
\value{
An S4 object with statistics for every DMR
\itemize{
   \item DMR - `dmrcate()`output, a S4 object with statistics for every DMR
   \item DMRRanges - `extractRanges()` output, a S4 object with Ranges from found DMRs
   \item annotation - `cpg.annotate()` output, a s4 object with statistics on every CpG site.
}
}
\description{
This function uses `limma` package to generate t-values. Then it smooths out these t-statistics using gaussian kernel smoothing.
The t-values that remain significant after smoothing and multiple testing, are grouped when they are within Lambda base pairs from each other.
Such a group is called a Differently Methylated Region (DMR).
}
