% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_QualityChecks.R
\name{FindBadSamples}
\alias{FindBadSamples}
\title{Title}
\usage{
FindBadSamples(x, badSamples = c(), n = 4, ylim = c(0, 5), showPDF = TRUE)
}
\arguments{
\item{showPDF}{}
}
\value{

}
\description{
Title
}
