% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPath.R
\name{getPath}
\alias{getPath}
\title{GetPath function}
\usage{
getPath(
  x,
  subName = "",
  subDir = "minfi_sets",
  extension = ".Rdata",
  mainDir = ".."
)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{subName}{Name of the file.}

\item{subDir}{Directory of the file. The default is the "minfi_sets" directory.}

\item{extension}{Extension of the file. Default is ".Rdata".}

\item{mainDir}{Main directory to save the file to. The default is the directory upstream from the working directory.}
}
\value{
Returns a filepath.
}
\description{
Function to automatically make a unique and descriptive path name based on the properties of the epigenetic set.
}
\examples{
getPath(list(name="test", params=list(combatChip = TRUE, combatRow=TRUE, recalibrate=FALSE)))
getPath("rgset", "plot3", "output")

}
