% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_Visualise.R
\name{MergeMetadata}
\alias{MergeMetadata}
\title{MergeMetadata function}
\usage{
MergeMetadata(x, df, by)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{df}{Metadata}

\item{by}{Name of the column to merge existing data with}
}
\value{
Updated Epigenetic set object.
}
\description{
Function to merge the metadata with the principal component data.
}
