% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_ProbeNormalisation.R
\name{ProbeBackgroundCheck}
\alias{ProbeBackgroundCheck}
\title{ProbeBackgroundCheck function}
\usage{
ProbeBackgroundCheck(x)
}
\arguments{
\item{x}{Epigenetic set object.}
}
\value{
an updated Epigenetic set object with the background check included.
}
\description{
A function that detects which probes from the rgset do not differ from the background probes.
}
