% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_ProbeNormalisation.R
\name{ProbeExclusion}
\alias{ProbeExclusion}
\title{ProbeExclusion function}
\usage{
ProbeExclusion(
  x,
  snps = c("CpG", "SBE"),
  sex = TRUE,
  background = TRUE,
  ExportExcludedProbes = TRUE,
  delay = FALSE
)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{snps}{Which SNPs position should considered when excluding probes on SNPs.}

\item{sex}{If TRUE, probes on the sex chromosomes are excluded.}

\item{background}{If TRUE, a background check will be completed and those probes which signals do not differ from the background noise, will be removed.}

\item{ExportExcludedProbes}{if TRUE, the locations of all the removed probes will be saved in the Epigenetic set and exported to a '.csv' file}

\item{delay}{If TRUE, the excluded probes will be saved, but not yet excluded.}
}
\value{
Updated Epigenetic set object.
}
\description{
Removes probes on SNPS, Sex chromosomes or that do not differ from the background noise.
}
