% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_Visualise.R
\name{BatchEffect_PCAHeatmap}
\alias{BatchEffect_PCAHeatmap}
\title{Title}
\usage{
BatchEffect_PCAHeatmap(x, Cat_vars = c(), Cont_vars = c())
}
\arguments{
\item{Cont_vars}{}
}
\value{

}
\description{
Title
}
