% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect_Visualise.R
\name{BatchEffect_Visualise}
\alias{BatchEffect_Visualise}
\title{Title}
\usage{
BatchEffect_Visualise(
  x,
  df,
  by,
  batch,
  Cat_vars = c(),
  Cont_vars = c(),
  n = 6,
  xlim = c(-1000, 2500),
  ylim = c(-1000, 1000)
)
}
\arguments{
\item{ylim}{}
}
\value{

}
\description{
Title
}
