% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_QualityChecks.R
\name{FindBadSamples}
\alias{FindBadSamples}
\title{FindBadSamples function}
\usage{
FindBadSamples(
  x,
  badSamples = c(),
  n = 4,
  ylim = c(0, 5),
  Swabname_position = 3
)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{badSamples}{Number or filenames of the bad idat files.}

\item{n}{number of density plots per plot. Max 4 is suggested.}

\item{ylim}{range of the y-axis}

\item{Swabname_position}{Position of swabname in the filename (separated by "_")}
}
\value{
Updated Epigenetic set object.
}
\description{
Function to visually find bad samples with non-bimodal distributions using density plots.
}
