% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_ProbeNormalisation.R
\name{ProbeExclusion}
\alias{ProbeExclusion}
\title{Title}
\usage{
ProbeExclusion(
  x,
  snps = c("CpG", "SBE"),
  sex = TRUE,
  background = TRUE,
  ExportSNPs = TRUE
)
}
\arguments{
\item{ExportSNPs}{}
}
\value{

}
\description{
Title
}
