% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_ProbeNormalisation.R
\name{ProbeNormalisation}
\alias{ProbeNormalisation}
\title{ProbeNormalisation function}
\usage{
ProbeNormalisation(x, quant = TRUE, save = FALSE)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{quant}{If TRUE, a quantile normalisation will be preformed, otherwise a functional normalisation will be done.}

\item{save}{If TRUE, the methylset will be saved to the "minfi_sets" sub-directory.}
}
\value{
Updated Epigenetic set object.
}
\description{
Functional or Quantile Normalisation of the CpG probes.
}
