% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_ProbeNormalisation.R
\name{ProbeNormalisation}
\alias{ProbeNormalisation}
\title{ProbeNormalisation function}
\usage{
ProbeNormalisation(x, method = "funnorm", nPCs = 3, save = FALSE, verbose = 0)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{method}{Type of normalisation algorithm. Options: funnorm, quantile, raw.}

\item{nPCs}{Number of principal components that funnorm uses.}

\item{save}{If TRUE, the methylset will be saved to the "minfi_sets" sub-directory.}

\item{verbose}{When set to 1 or higher more messages will be shown.}
}
\value{
Updated Epigenetic set object.
}
\description{
Functional or Quantile Normalisation of the CpG probes.
}
