% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_QualityChecks.R
\name{RemoveBadSamples}
\alias{RemoveBadSamples}
\title{RemoveBadSamples function}
\usage{
RemoveBadSamples(
  x,
  badSamplesNames = c(),
  badSamplesNum = c(),
  Swabname_position = 3
)
}
\arguments{
\item{x}{Epigenetic set object.}

\item{badSamplesNames}{Filenames from bad samples}

\item{badSamplesNum}{Index numbers from bad samples.}

\item{Swabname_position}{Position of swabname in the filename (separated by "_")}
}
\value{
Updated Epigenetic set object.
}
\description{
Function to remove bad samples from an epigenetic set. This function also outputs a pdf with density plots highlighting the removed samples and without the bad samples.
}
