% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_QualityChecks.R
\name{RemoveBadSamples}
\alias{RemoveBadSamples}
\title{Title}
\usage{
RemoveBadSamples(x, badSamplesNames = c(), badSamplesNum = c())
}
\arguments{
\item{badSamplesNum}{}
}
\value{

}
\description{
Title
}
